+++ lib/features/home/presentation/page/home_page.dart
@@ -7,7 +7,7 @@ import 'package:technical_support_artphoto/features/home/presentation/widgets/gr
 import 'package:technical_support_artphoto/features/home/presentation/widgets/grid_view_home_page.dart';
 import 'package:technical_support_artphoto/features/home/presentation/widgets/my_custom_refresh_indicator.dart';
 import 'package:technical_support_artphoto/features/technics/presentation/page/technic_add.dart';
-import '../widgets/title_appbar_homepage.dart';
+import '../widgets/appbar_homepage.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({super.key});
@@ -25,12 +25,7 @@ class _HomePageState extends State<HomePage> {
     final providerModel = Provider.of<ProviderModel>(context);
     final double height = MediaQuery.sizeOf(context).height;
     return Scaffold(
-      appBar: AppBar(
-          title: TitleAppbarHomepage()
-          // actions: [
-          //   MenuHomePage(),
-          // ],
-          ),
+      appBar: AppBarHomepage(),
       body: SafeArea(
         child: WarpIndicator(
             controller: _controller,
